# backend/chat_service/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.ChatListResponse:
    properties:
      chats:
        items:
          $ref: '#/definitions/handlers.ChatResponse'
        type: array
      total:
        description: Optional, if we implement count
        type: integer
    type: object
  handlers.ChatResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
  handlers.CreateChatRequest:
    properties:
      my_encrypted_key:
        type: string
      participant_encrypted_key:
        type: string
      participant_id:
        type: string
    required:
    - my_encrypted_key
    - participant_encrypted_key
    - participant_id
    type: object
  handlers.EncryptedKeyResponse:
    properties:
      encrypted_key:
        type: string
    type: object
  handlers.IsMemberResponse:
    properties:
      is_member:
        type: boolean
    type: object
host: localhost:8083
info:
  contact: {}
  description: API Server for Chat Service
  title: Chat Service API
  version: "1.0"
paths:
  /chats:
    get:
      description: Get a list of chats for the current user with pagination
      parameters:
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ChatListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get user chats
      tags:
      - chats
    post:
      consumes:
      - application/json
      description: Creates a new one-to-one chat between the current user and a participant
      parameters:
      - description: Chat creation parameters
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateChatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.ChatResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a new chat
      tags:
      - chats
  /chats/{id}:
    get:
      description: Get details of a specific chat
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ChatResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get chat by ID
      tags:
      - chats
  /chats/{id}/key:
    get:
      description: Get the encrypted AES key for the current user in the chat
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.EncryptedKeyResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get encrypted key
      tags:
      - chats
  /chats/{id}/member:
    get:
      description: Check if the current user is a member of the chat
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.IsMemberResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Check membership
      tags:
      - chats
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"